fix: code quality issues from Task 1 review
- MySQLConnectPool.cpp: Fix stop() deadlock by releasing lock before joining thread
- MySQLConnectPool.h: Change pool_size_ from int to size_t
- RedisConnectPool: Store password as member variable for future reconnection
- ConfigManager.cpp: Mask password fields in debug config output
- vehicle_rental.sql: Add comments explaining missing FKs on order_no and
  noting seed passwords are plaintext for demo purposes
- Defer.h: Make move-only (delete copy ctor/assignment, add move ops)
- Singleton.h: Remove unused includes (memory, mutex, iostream)
- Global.h: Remove unused includes (memory, mutex, iostream, json),
  change MSG_IDS to scoped enum class MsgId
- Logger.cpp: Rename time_t local variable to tt to avoid shadowing
- RedisManager.cpp: Read Redis pool size from config instead of hardcoding

+++ GateServer/MySQLConnectPool.h
@@ -46,7 +46,7 @@ private:
     std::string password_;
     std::string dbName_;
     std::string user_;
-    int pool_size_;
+    size_t pool_size_;
     std::mutex mutex_;
     std::queue<std::unique_ptr<SqlConnection>> connections_;
     std::atomic<bool> is_running_;
